Update readme todo list
Remove `- [ ] Add ability for elements created after pageload to trigger events.` from readme todo list.

While it sounds great, it's not practical for a jquery plugin like this that needs to know the event types in advance in order to know what to listen for. If those event types, by design, are ascertained by looking at the `data` attributes of an element, it's kind of impossible to delegate events _before_ an element exists in the DOM.

Instead, the proper workflow here would be to re-initialize jquery track when any new elements are injected into the document. i.e. 

```js
$.get('/some/api/url', function (data) {
  // use data to render something in the DOM
  // $('<selector to target new DOM elements>').track();
});
```